cifar-10/requirements.txt
numpy
pytorch
scikit-learn
tensorboardX
torchvision
tqdm
